Intent: Consolidate numeric helpers under utils Operators involved: [Coherence, Coupling, Transition] Affected invariants: [#8, #9] Key changes: - Move numeric helper implementations to tnfr.utils.numeric with updated exports - Rework tnfr.helpers package as a deprecated shim emitting DeprecationWarning - Redirect internal, test, and doc imports to tnfr.utils while documenting the migration Expected risks/dissonances: Compatibility callers may miss the deprecation warning; shim preserves behaviour to limit dissonance Metrics: pytest structural/dynamics subsets (skipped: missing numpy) will continue to validate numeric helpers once numpy is present Equivalence map: tnfr.helpers.numeric.* -> tnfr.utils.numeric.*, tnfr.helpers.* -> tnfr.utils.*
