+++ src/tnfr/dynamics/coordination.py
@@ -19,7 +19,7 @@ from ..constants import (
     normalise_state_token,
 )
 from ..glyph_history import append_metric
-from ..helpers.numeric import angle_diff
+from ..utils.numeric import angle_diff
 from ..metrics.common import ensure_neighbors_map
 from ..metrics.trig import neighbor_phase_mean_list
 from ..metrics.trig_cache import get_trig_cache
