+++ src/tnfr/helpers/numeric.py
@@ -1,144 +1,17 @@
-"""Numeric helper functions and compensated summation utilities."""
+"""Compatibility layer for :mod:`tnfr.helpers.numeric`."""
 
 from __future__ import annotations
 
-import math
-from collections.abc import Iterable, Sequence
-from typing import Any
+import warnings
 
-__all__ = (
-    "clamp",
-    "clamp01",
-    "within_range",
-    "similarity_abs",
-    "kahan_sum_nd",
-    "angle_diff",
-    "angle_diff_array",
+from ..utils.numeric import *  # noqa: F401,F403 - re-export compatibility
+from ..utils.numeric import __all__ as _UTILS_ALL
+
+warnings.warn(
+    "'tnfr.helpers.numeric' is deprecated and will be removed in a future release; "
+    "import from 'tnfr.utils.numeric' instead.",
+    DeprecationWarning,
+    stacklevel=2,
 )
 
-
-def clamp(x: float, a: float, b: float) -> float:
-    """Return ``x`` clamped to the ``[a, b]`` interval."""
-    return max(a, min(b, x))
-
-
-def clamp01(x: float) -> float:
-    """Clamp ``x`` to the ``[0,1]`` interval."""
-    return clamp(float(x), 0.0, 1.0)
-
-
-def within_range(val: float, lower: float, upper: float, tol: float = 1e-9) -> bool:
-    """Return ``True`` if ``val`` lies in ``[lower, upper]`` within ``tol``.
-
-    The comparison uses absolute differences instead of :func:`math.isclose`.
-    """
-
-    v = float(val)
-    return lower <= v <= upper or abs(v - lower) <= tol or abs(v - upper) <= tol
-
-
-def _norm01(x: float, lo: float, hi: float) -> float:
-    """Normalize ``x`` to the unit interval given bounds.
-
-    ``lo`` and ``hi`` delimit the original value range. When ``hi`` is not
-    greater than ``lo`` the function returns ``0.0`` to avoid division by
-    zero. The result is clamped to ``[0,1]``.
-    """
-
-    if hi <= lo:
-        return 0.0
-    return clamp01((float(x) - float(lo)) / (float(hi) - float(lo)))
-
-
-def similarity_abs(a: float, b: float, lo: float, hi: float) -> float:
-    """Return absolute similarity of ``a`` and ``b`` over ``[lo, hi]``.
-
-    It computes ``1`` minus the normalized absolute difference between
-    ``a`` and ``b``. Values are scaled using :func:`_norm01` so the result
-    falls within ``[0,1]``.
-    """
-
-    return 1.0 - _norm01(abs(float(a) - float(b)), 0.0, hi - lo)
-
-
-def kahan_sum_nd(values: Iterable[Sequence[float]], dims: int) -> tuple[float, ...]:
-    """Return compensated sums of ``values`` with ``dims`` components.
-
-    Each component of the tuples in ``values`` is summed independently using the
-    Kahan–Babuška (Neumaier) algorithm to reduce floating point error.
-    """
-    if dims < 1:
-        raise ValueError("dims must be >= 1")
-    totals = [0.0] * dims
-    comps = [0.0] * dims
-    for vs in values:
-        for i in range(dims):
-            v = vs[i]
-            t = totals[i] + v
-            if abs(totals[i]) >= abs(v):
-                comps[i] += (totals[i] - t) + v
-            else:
-                comps[i] += (v - t) + totals[i]
-            totals[i] = t
-    return tuple(float(totals[i] + comps[i]) for i in range(dims))
-
-
-def angle_diff(a: float, b: float) -> float:
-    """Return the minimal difference between two angles in radians."""
-    return (float(a) - float(b) + math.pi) % math.tau - math.pi
-
-
-def angle_diff_array(
-    a: Sequence[float] | "np.ndarray",
-    b: Sequence[float] | "np.ndarray",
-    *,
-    np: Any,
-    out: "np.ndarray | None" = None,
-    where: "np.ndarray | None" = None,
-) -> "np.ndarray":
-    """Vectorised :func:`angle_diff` compatible with NumPy arrays.
-
-    Parameters
-    ----------
-    a, b:
-        Arrays or array-like sequences containing the minuend and subtrahend
-        angles respectively.
-    np:
-        NumPy module used to materialise the vectorised operations. Passing the
-        module explicitly keeps :mod:`tnfr.helpers.numeric` free from a hard
-        dependency on NumPy while still allowing accelerated paths to reuse the
-        canonical angular difference implementation.
-    out:
-        Optional buffer receiving the computed differences. When provided the
-        buffer is updated in-place; otherwise a new array is returned.
-    where:
-        Optional boolean mask restricting the positions that are updated. When
-        omitted all entries are processed.
-
-    Returns
-    -------
-    numpy.ndarray
-        Array containing the wrapped angle differences ``(a - b)`` normalised
-        to ``[-π, π]``. The result shares the provided output buffer when
-        ``out`` is supplied.
-    """
-
-    if np is None:
-        raise TypeError("angle_diff_array requires a NumPy module")
-
-    kwargs = {"where": where} if where is not None else {}
-    minuend = np.asarray(a, dtype=float)
-    subtrahend = np.asarray(b, dtype=float)
-    if out is None:
-        out = np.empty_like(minuend, dtype=float)
-        if where is not None:
-            out.fill(0.0)
-    else:
-        if getattr(out, "shape", None) != minuend.shape:
-            raise ValueError("out must match the broadcasted shape of inputs")
-
-    np.subtract(minuend, subtrahend, out=out, **kwargs)
-    np.add(out, math.pi, out=out, **kwargs)
-    np.mod(out, math.tau, out=out, **kwargs)
-    np.subtract(out, math.pi, out=out, **kwargs)
-    return out
+__all__ = _UTILS_ALL
