+++ src/tnfr/utils/__init__.py
@@ -47,6 +47,15 @@ from .graph import (
     mark_dnfr_prep_dirty,
     supports_add_edge,
 )
+from .numeric import (
+    angle_diff,
+    angle_diff_array,
+    clamp,
+    clamp01,
+    kahan_sum_nd,
+    similarity_abs,
+    within_range,
+)
 from .io import (
     DEFAULT_PARAMS,
     JsonDumpsParams,
@@ -76,6 +85,10 @@ __all__ = (
     "MAX_MATERIALIZE_DEFAULT",
     "negative_weights_warn_once",
     "mix_groups",
+    "angle_diff",
+    "angle_diff_array",
+    "clamp",
+    "clamp01",
     "auto_chunk_size",
     "resolve_chunk_size",
     "CacheManager",
@@ -106,6 +119,9 @@ __all__ = (
     "DEFAULT_PARAMS",
     "json_dumps",
     "clear_orjson_param_warnings",
+    "kahan_sum_nd",
+    "similarity_abs",
+    "within_range",
     "validate_window",
     "run_validators",
     "_configure_root",
